requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
asttokens
attrs
Authlib
backoff
bcrypt
beautifulsoup4
build
cachetools
certifi
cffi
chardet
charset-normalizer
chromadb
click
colorama
coloredlogs
comm
cryptography
datasets
debugpy
decorator
deep-translator
deprecation
dill
distro
durationpy
executing
fastapi
ffmpeg-python
filelock
flatbuffers
frozenlist
fsspec
ftfy
future
google-auth
googleapis-common-protos
googletrans
grpcio
h11
h2
hf-xet
hpack
hstspreload
httpcore
httptools
httpx
huggingface-hub
humanfriendly
hyperframe
idna
imagebind
importlib_metadata
importlib_resources
ipykernel
ipython
ipython_pygments_lexers
jedi
Jinja2
jiter
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
kubernetes
langchain-core
langchain-openai
langdetect
langsmith
llvmlite
markdown-it-py
MarkupSafe
matplotlib-inline
mdurl
mmh3
more-itertools
mpmath
multidict
multiprocess
nest-asyncio
networkx
numba
numpy
oauthlib
onnxruntime
open_clip_torch
openai
openai-whisper
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
orjson
overrides
packaging
pandas
parso
pillow
platformdirs
posthog
prompt_toolkit
propcache
protobuf
psutil
pure_eval
pyarrow
pyasn1
pyasn1_modules
pybase64
pycparser
pydantic
pydantic_core
Pygments
PyPika
pyproject_hooks
pyreadline3
python-dateutil
python-dotenv
python-multipart
pytz
PyYAML
pyzmq
referencing
regex
requests
requests-oauthlib
requests-toolbelt
rfc3986
rich
rpds-py
rsa
safetensors
sentencepiece
setuptools
shellingham
six
sniffio
soundfile
soupsieve
stack-data
starlette
sympy
tenacity
tiktoken
timm
tokenizers
torch
torchaudio
torchcodec
torchvision
tornado
tqdm
traitlets
transformers
typer
typing-inspection
typing_extensions
tzdata
urllib3
uvicorn
validators
watchfiles
wcwidth
weaviate-client
websocket-client
websockets
xxhash
yarl
zipp
zstandard
